trip controller logic
